pyproject.toml
[tool.poetry]
name = "repl_python3_beta_pythonbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
replit = "^3.2.4"
pytz = "^2021.1"
python-telegram-bot = "^13.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
